Extract bugsnag service into client/app/services/ bugsnag.js
APERTA-8110
